+++ scripts/run_checks.py
@@ -17,7 +17,9 @@ COMPILE_TARGETS = (
     "goldmonitor/review_notes.py",
     "goldmonitor/scheduler.py",
     "goldmonitor/socket_alert_rules.py",
+    "goldmonitor/socket_history_review.py",
     "goldmonitor/socket_portfolio.py",
+    "goldmonitor/socket_risk_analysis.py",
     "tests/risk_contract_check.py",
     "tests/update_logic_check.py",
     "tests/startup_contract_check.py",
