+++ static/app.js
@@ -123,22 +123,6 @@ let pendingPortfolioUndoMessage = '';
 let portfolioImportPreview = null;
 let portfolioImportPreviewRequestSeq = 0;
 let activeWatchTargetId = null;
-let historyView = 'prices';
-let eventTimelineState = { events: [], summary: {}, filters: {}, range: {}, price_summary: {} };
-let eventTimelineRange = 60;
-let eventTimelineTypes = ['price_summary', 'alert', 'risk_analysis', 'news', 'data_status', 'review_note'];
-let selectedTimelineEventId = null;
-let pendingTimelineFocus = null;
-let reviewNoteEditorState = { id: '', related_event_id: '', related_event_type: '', related_event_title: '' };
-let reviewNotesRefreshTimer = null;
-const EVENT_TIMELINE_TYPE_DEFS = [
-  { type: 'price_summary', label: '价格摘要' },
-  { type: 'alert', label: '预警' },
-  { type: 'risk_analysis', label: '风险分析' },
-  { type: 'news', label: '新闻' },
-  { type: 'data_status', label: '数据状态' },
-  { type: 'review_note', label: '复盘笔记' },
-];
 let appSettings = {
   onboarding_started: false,
   onboarding_completed: false,
@@ -245,12 +229,7 @@ let alertEntries = [];
 let alertLogSearch = '';
 let activeAlert = null;
 let mergedAlertCount = 0;
-let riskAnalysisRunning = false;
-let riskAnalysisHistory = [];
-let riskComparisonSelection = [];
-let pendingRiskForceTrigger = null;
 let deepseekModelOptions = ['deepseek-v4-pro', 'deepseek-v4-flash', 'deepseek-chat', 'deepseek-reasoner'];
-let latestPriceHistoryState = { items: [], stats: {}, total: 0 };
 let latestSourceHealthState = { items: [], summary: {} };
 let latestSourceComparisonState = { items: [], summary: {}, status: 'insufficient' };
 let dailyDigestStatusState = {};
@@ -967,96 +946,7 @@ socket.on('update_status', data => {
   applyUpdateStatus(data || {});
 });
 
-socket.on('risk_analysis_status', data => {
-  riskAnalysisRunning = !!(data && data.running);
-  if (data && data.message) {
-    applyRiskStatus(data.message, riskAnalysisRunning ? 'loading' : '');
-  }
-  updateRiskButtonState();
-});
-
-socket.on('risk_analysis_result', data => {
-  riskAnalysisRunning = false;
-  const usageText = formatRiskUsage(data && data.usage ? data.usage : null);
-  renderRiskDiagnostic(null);
-  applyRiskStatus(usageText ? '分析完成。' + usageText : '分析完成。', '');
-  renderRiskSnapshot(data && data.snapshot ? data.snapshot : null);
-  renderRiskStructured(data && data.structured ? data.structured : null);
-  document.getElementById('riskResult').textContent = data && data.content ? data.content : '未返回分析内容。';
-  pendingRiskForceTrigger = null;
-  setRiskForceButtonVisible(false);
-  if (data && data.snapshot) {
-    addChartEvent({
-      type: 'risk',
-      level: 'analysis',
-      timestamp: data.snapshot.analysis_time || '',
-      label: data.structured && data.structured.risk_level ? '风险 ' + data.structured.risk_level : '风险分析',
-      message: data.content || '',
-    });
-  }
-  updateRiskButtonState();
-});
-
-socket.on('risk_analysis_error', data => {
-  riskAnalysisRunning = false;
-  applyRiskStatus(data && data.message ? data.message : '风险分析失败。', 'error');
-  renderRiskDiagnostic(data && data.diagnostic ? data.diagnostic : null);
-  if (data && data.snapshot) renderRiskSnapshot(data.snapshot);
-  updateRiskButtonState();
-});
-
-socket.on('risk_analysis_cache_hit', data => {
-  riskAnalysisRunning = false;
-  renderRiskDiagnostic(null);
-  const age = data && data.cache_age_seconds != null ? Math.max(0, Number(data.cache_age_seconds)) : 0;
-  const ageText = age >= 60 ? Math.floor(age / 60) + ' 分钟前' : Math.max(1, age) + ' 秒前';
-  applyRiskStatus((data && data.message ? data.message : '已复用最近同一行情分析。') + ' 生成于 ' + ageText + '。', '');
-  renderRiskSnapshot(data && data.snapshot ? data.snapshot : null);
-  renderRiskStructured(data && data.structured ? data.structured : null);
-  document.getElementById('riskResult').textContent = data && data.content ? data.content : '暂无缓存分析内容。';
-  pendingRiskForceTrigger = data && data.trigger ? data.trigger : null;
-  setRiskForceButtonVisible(true);
-  if (data && data.snapshot) {
-    addChartEvent({
-      type: 'risk',
-      level: 'analysis',
-      timestamp: data.snapshot.analysis_time || '',
-      label: data.structured && data.structured.risk_level ? '风险 ' + data.structured.risk_level : '风险分析',
-      message: data.content || '',
-    });
-  }
-  updateRiskButtonState();
-});
-
-socket.on('risk_analysis_history_updated', data => {
-  applyRiskHistory(data || {});
-});
-
-socket.on('risk_model_test_result', data => {
-  const el = document.getElementById('riskModelTestStatus');
-  if (!el) return;
-  el.textContent = data && data.message ? data.message : '模型测试完成。';
-  el.className = 'model-test-status ' + (data && data.ok ? 'ok' : 'fail');
-});
-
-socket.on('open_risk_analysis', data => {
-  if (!openRiskAnalysis()) return;
-  if (data && data.run) requestRiskAnalysis({ source: data.source || 'floating_price' });
-});
-
-socket.on('risk_model_options_updated', data => {
-  if (!data || data.provider !== 'deepseek') return;
-  deepseekModelOptions = Array.isArray(data.models) && data.models.length ? data.models : deepseekModelOptions;
-  const currentModel = document.getElementById('setDeepseekModel').value || appSettings.deepseek_model || 'deepseek-v4-pro';
-  renderDeepseekModelOptions(currentModel);
-  const status = document.getElementById('deepseekModelStatus');
-  if (status) {
-    status.textContent = data.error
-      ? data.error + ' 当前显示兜底模型。'
-      : '模型列表已更新，来源：' + (data.source === 'api' ? '接口' : '兜底列表') + '。';
-    status.dataset.state = data.error ? 'error' : 'ok';
-  }
-});
+registerRiskAnalysisSocketHandlers(socket);
 
 socket.on('source_health_updated', data => {
   renderSourceHealth(data || {});
@@ -1077,73 +967,7 @@ socket.on('market_source_retry_result', data => {
   if (data && data.source_health) renderSourceHealth(data.source_health);
 });
 
-socket.on('price_history_updated', data => {
-  applyPriceHistory(data || {});
-});
-
-socket.on('price_history_export_ready', data => {
-  if (!data || !data.content) return;
-  downloadText(data.filename || 'GoldMonitor-price-history.csv', data.content, 'text/csv;charset=utf-8');
-});
-
-socket.on('event_timeline_updated', data => {
-  applyEventTimeline(data || {});
-});
-
-socket.on('event_timeline_error', data => {
-  setTimelineStatus((data && data.message) || '事件时间轴加载失败。', 'fail');
-});
-
-socket.on('review_note_saved', data => {
-  setReviewNoteSaving(false);
-  if (data && data.ok === false) {
-    setReviewNoteEditorStatus(data.message || '复盘笔记保存失败。', 'fail');
-    return;
-  }
-  const note = data && (data.note || data.item) || {};
-  if (note.id) {
-    pendingTimelineFocus = {
-      type: 'review_note',
-      sourceId: String(note.id),
-      timestamp: note.timestamp || '',
-    };
-  }
-  closeReviewNoteEditor();
-  setTimelineStatus((data && data.message) || '复盘笔记已保存。', 'ok');
-  queueReviewNotesTimelineRefresh();
-});
-
-socket.on('review_note_deleted', data => {
-  if (data && data.ok === false) {
-    setTimelineStatus(data.message || '复盘笔记删除失败。', 'fail');
-    return;
-  }
-  selectedTimelineEventId = null;
-  closeReviewNoteEditor();
-  setTimelineStatus((data && data.message) || '复盘笔记已删除。', 'ok');
-  queueReviewNotesTimelineRefresh();
-});
-
-socket.on('review_note_error', data => {
-  setReviewNoteSaving(false);
-  const message = (data && data.message) || '复盘笔记操作失败。';
-  const editor = document.getElementById('reviewNoteEditor');
-  if (editor && !editor.hidden) setReviewNoteEditorStatus(message, 'fail');
-  else setTimelineStatus(message, 'fail');
-});
-
-socket.on('review_notes_updated', () => {
-  queueReviewNotesTimelineRefresh();
-});
-
-socket.on('review_report_exported', data => {
-  const count = data && Number.isFinite(Number(data.count)) ? Number(data.count) : 0;
-  setTimelineStatus(data && data.saved_path ? '已导出 ' + count + ' 条事件，保存至 ' + data.saved_path : '复盘报告已导出。', 'ok');
-});
-
-socket.on('review_report_error', data => {
-  setTimelineStatus((data && data.message) || '复盘报告导出失败。', 'fail');
-});
+registerHistoryReviewSocketHandlers(socket);
 
 socket.on('config_backup_ready', data => {
   if (!data) return;
@@ -1547,572 +1371,6 @@ function refreshSourceHealth() {
   socket.emit('get_source_health');
 }
 
-function applyPriceHistory(data) {
-  if (Array.isArray(data.events)) {
-    chartEvents = normalizeChartEvents(data.events);
-  }
-  if (data && data.scope === 'chart') {
-    chartHistoryState = Object.assign({ period: data.period || chartPeriod, items: [] }, data);
-    if (chartHistoryState.period === chartPeriod) switchChartData();
-    return;
-  }
-  latestPriceHistoryState = Object.assign({ items: [], stats: {}, total: 0 }, data || {});
-  renderHistory(latestPriceHistoryState);
-}
-
-function historyStatValue(stats, field) {
-  const item = stats && stats.rmb ? stats.rmb[field] : null;
-  if (item == null) return '--';
-  return Number(item).toLocaleString('en-US', { maximumFractionDigits: 2 });
-}
-
-function renderHistory(data) {
-  const statsEl = document.getElementById('historyStats');
-  const listEl = document.getElementById('historyList');
-  if (!statsEl || !listEl) return;
-  const stats = data.stats || {};
-  const rmb = stats.rmb || {};
-  const items = Array.isArray(data.items) ? data.items : [];
-  const statItems = [
-    ['样本', data.total || items.length || 0],
-    ['最高 RMB', rmb.high == null ? '--' : '¥' + historyStatValue(stats, 'high')],
-    ['最低 RMB', rmb.low == null ? '--' : '¥' + historyStatValue(stats, 'low')],
-    ['变动', rmb.change_pct == null ? '--' : Number(rmb.change_pct).toFixed(2) + '%'],
-  ];
-  statsEl.innerHTML = statItems.map(item => (
-    '<div class="history-stat"><div class="history-stat-label">' + escapeHtml(item[0]) + '</div><div class="history-stat-value">' + escapeHtml(item[1]) + '</div></div>'
-  )).join('');
-  if (!items.length) {
-    listEl.innerHTML = '<div class="history-empty">暂无历史数据</div>';
-    return;
-  }
-  const rows = items.slice(-240).reverse().map(item => [
-    '<div class="history-row">',
-    '<span>' + escapeHtml((item.timestamp || '').replace('T', ' ')) + '</span>',
-    '<span>' + escapeHtml(item.rmb == null ? '--' : '¥' + Number(item.rmb).toFixed(2)) + '</span>',
-    '<span>' + escapeHtml(item.usd == null ? '--' : '$' + Number(item.usd).toFixed(2)) + '</span>',
-    '<span>' + escapeHtml(item.rate == null ? '--' : Number(item.rate).toFixed(4)) + '</span>',
-    '</div>',
-  ].join('')).join('');
-  listEl.innerHTML = '<div class="history-row"><span>时间</span><span>RMB/克</span><span>USD/oz</span><span>汇率</span></div>' + rows;
-}
-
-function openHistory() {
-  document.getElementById('historyBackdrop').classList.add('show');
-  renderTimelineTypeFilters();
-  switchHistoryView(historyView || 'prices', true);
-  refreshHistory();
-}
-
-function closeHistory() {
-  document.getElementById('historyBackdrop').classList.remove('show');
-  closeReviewNoteEditor();
-}
-
-function onHistoryBackdrop(event) {
-  if (event.target.id === 'historyBackdrop') closeHistory();
-}
-
-function refreshHistory() {
-  socket.emit('get_price_history', { limit: 600 });
-}
-
-function switchHistoryView(view, skipRefresh) {
-  historyView = view === 'timeline' ? 'timeline' : 'prices';
-  const isTimeline = historyView === 'timeline';
-  const priceTab = document.getElementById('historyTabPrices');
-  const timelineTab = document.getElementById('historyTabTimeline');
-  const pricePanel = document.getElementById('historyPanelPrices');
-  const timelinePanel = document.getElementById('historyPanelTimeline');
-  if (priceTab) {
-    priceTab.classList.toggle('active', !isTimeline);
-    priceTab.setAttribute('aria-selected', String(!isTimeline));
-  }
-  if (timelineTab) {
-    timelineTab.classList.toggle('active', isTimeline);
-    timelineTab.setAttribute('aria-selected', String(isTimeline));
-  }
-  if (pricePanel) pricePanel.classList.toggle('active', !isTimeline);
-  if (timelinePanel) timelinePanel.classList.toggle('active', isTimeline);
-  const csvBtn = document.getElementById('exportHistoryCsvButton');
-  const reportBtn = document.getElementById('exportReviewReportButton');
-  if (csvBtn) csvBtn.style.display = isTimeline ? 'none' : '';
-  if (reportBtn) reportBtn.style.display = isTimeline ? '' : 'none';
-  if (isTimeline && !skipRefresh) refreshEventTimeline();
-}
-
-function refreshHistoryCurrentView() {
-  if (historyView === 'timeline') {
-    refreshEventTimeline();
-    return;
-  }
-  refreshHistory();
-}
-
-function timelineTypeLabel(type) {
-  const found = EVENT_TIMELINE_TYPE_DEFS.find(item => item.type === type);
-  return found ? found.label : type;
-}
-
-function selectedTimelineEvent() {
-  const events = Array.isArray(eventTimelineState.events) ? eventTimelineState.events : [];
-  return events.find(item => item.id === selectedTimelineEventId) || null;
-}
-
-function reviewNoteIdFromEvent(event) {
-  if (!event) return '';
-  const payload = event.payload && typeof event.payload === 'object' ? event.payload : {};
-  return String(payload.note_id || payload.id || event.note_id || event.id || '');
-}
-
-function reviewNoteLocalInputValue(value) {
-  const parsed = timelineDateFromValue(value) || new Date();
-  const pad = number => String(number).padStart(2, '0');
-  return [
-    parsed.getFullYear(),
-    pad(parsed.getMonth() + 1),
-    pad(parsed.getDate()),
-  ].join('-') + 'T' + pad(parsed.getHours()) + ':' + pad(parsed.getMinutes());
-}
-
-function setReviewNoteEditorStatus(message, type) {
-  const el = document.getElementById('reviewNoteEditorStatus');
-  if (!el) return;
-  el.textContent = message || '';
-  el.className = 'review-note-editor-status' + (type ? ' ' + type : '');
-}
-
-function setReviewNoteSaving(saving) {
-  const button = document.getElementById('saveReviewNoteButton');
-  if (!button) return;
-  button.disabled = Boolean(saving);
-  button.textContent = saving ? '正在保存' : '保存笔记';
-}
-
-function setReviewNoteEditorRelation(state) {
-  const relation = document.getElementById('reviewNoteRelation');
-  if (!relation) return;
-  if (!state.related_event_id) {
-    relation.textContent = '独立笔记';
-    return;
-  }
-  relation.textContent = '关联：' + timelineTypeLabel(state.related_event_type) + ' · ' + (state.related_event_title || state.related_event_id);
-}
-
-function showReviewNoteEditor(options) {
-  const state = Object.assign({
-    id: '',
-    timestamp: '',
-    title: '',
-    content: '',
-    related_event_id: '',
-    related_event_type: '',
-    related_event_title: '',
-  }, options || {});
-  reviewNoteEditorState = {
-    id: String(state.id || ''),
-    related_event_id: String(state.related_event_id || ''),
-    related_event_type: String(state.related_event_type || ''),
-    related_event_title: String(state.related_event_title || ''),
-  };
-  const editor = document.getElementById('reviewNoteEditor');
-  const heading = document.getElementById('reviewNoteEditorHeading');
-  const timestamp = document.getElementById('reviewNoteTimestamp');
-  const title = document.getElementById('reviewNoteTitle');
-  const content = document.getElementById('reviewNoteContent');
-  if (!editor || !timestamp || !title || !content) return;
-  if (heading) heading.textContent = state.id ? '编辑复盘笔记' : '新增复盘笔记';
-  timestamp.value = reviewNoteLocalInputValue(state.timestamp);
-  title.value = state.title || '';
-  content.value = state.content || '';
-  setReviewNoteEditorRelation(reviewNoteEditorState);
-  setReviewNoteEditorStatus('', '');
-  setReviewNoteSaving(false);
-  editor.hidden = false;
-  requestAnimationFrame(() => content.focus());
-}
-
-function openReviewNoteEditor() {
-  showReviewNoteEditor({ timestamp: new Date() });
-}
-
-function openReviewNoteEditorFromSelectedEvent() {
-  const event = selectedTimelineEvent();
-  if (!event || event.type === 'review_note') {
-    setTimelineStatus('请先选择一条行情、预警、分析、新闻或数据事件。', 'fail');
-    return;
-  }
-  showReviewNoteEditor({
-    timestamp: event.timestamp || new Date(),
-    related_event_id: event.id,
-    related_event_type: event.type,
-    related_event_title: event.title || timelineTypeLabel(event.type),
-  });
-}
-
-function editSelectedReviewNote() {
-  const event = selectedTimelineEvent();
-  if (!event || event.type !== 'review_note') return;
-  const payload = event.payload && typeof event.payload === 'object' ? event.payload : {};
-  showReviewNoteEditor({
-    id: reviewNoteIdFromEvent(event),
-    timestamp: payload.timestamp || event.timestamp,
-    title: payload.title || event.title || '',
-    content: payload.content || event.summary || '',
-    related_event_id: payload.related_event_id || '',
-    related_event_type: payload.related_event_type || '',
-    related_event_title: payload.related_event_title || '',
-  });
-}
-
-function closeReviewNoteEditor() {
-  const editor = document.getElementById('reviewNoteEditor');
-  if (editor) editor.hidden = true;
-  reviewNoteEditorState = { id: '', related_event_id: '', related_event_type: '', related_event_title: '' };
-  setReviewNoteEditorStatus('', '');
-  setReviewNoteSaving(false);
-}
-
-function saveReviewNote() {
-  const timestamp = document.getElementById('reviewNoteTimestamp');
-  const title = document.getElementById('reviewNoteTitle');
-  const content = document.getElementById('reviewNoteContent');
-  if (!timestamp || !title || !content) return;
-  const payload = {
-    timestamp: timestamp.value.trim(),
-    title: title.value.trim(),
-    content: content.value.trim(),
-    related_event_id: reviewNoteEditorState.related_event_id,
-    related_event_type: reviewNoteEditorState.related_event_type,
-    related_event_title: reviewNoteEditorState.related_event_title,
-  };
-  if (!payload.timestamp) {
-    setReviewNoteEditorStatus('请选择笔记时间。', 'fail');
-    timestamp.focus();
-    return;
-  }
-  if (!payload.content) {
-    setReviewNoteEditorStatus('请输入笔记内容。', 'fail');
-    content.focus();
-    return;
-  }
-  if (payload.title.length > 80 || payload.content.length > 2000) {
-    setReviewNoteEditorStatus('标题最多 80 个字符，内容最多 2000 个字符。', 'fail');
-    return;
-  }
-  if (reviewNoteEditorState.id) payload.id = reviewNoteEditorState.id;
-  setReviewNoteSaving(true);
-  setReviewNoteEditorStatus('正在保存复盘笔记...', '');
-  socket.emit('save_review_note', payload);
-}
-
-function deleteSelectedReviewNote() {
-  const event = selectedTimelineEvent();
-  if (!event || event.type !== 'review_note') return;
-  const noteId = reviewNoteIdFromEvent(event);
-  if (!noteId) {
-    setTimelineStatus('缺少笔记标识，无法删除。', 'fail');
-    return;
-  }
-  if (!window.confirm('确定删除复盘笔记“' + (event.title || '未命名笔记') + '”？')) return;
-  setTimelineStatus('正在删除复盘笔记...', '');
-  socket.emit('delete_review_note', { id: noteId });
-}
-
-function queueReviewNotesTimelineRefresh() {
-  if (reviewNotesRefreshTimer) clearTimeout(reviewNotesRefreshTimer);
-  reviewNotesRefreshTimer = setTimeout(() => {
-    reviewNotesRefreshTimer = null;
-    const backdrop = document.getElementById('historyBackdrop');
-    if (historyView === 'timeline' && backdrop && backdrop.classList.contains('show')) refreshEventTimeline();
-  }, 80);
-}
-
-function renderTimelineTypeFilters() {
-  const box = document.getElementById('timelineTypeFilters');
-  if (!box) return;
-  box.innerHTML = EVENT_TIMELINE_TYPE_DEFS.map(item => {
-    const checked = eventTimelineTypes.includes(item.type) ? ' checked' : '';
-    return [
-      '<label>',
-      '<input type="checkbox" value="' + escapeHtml(item.type) + '"' + checked + ' onchange="toggleTimelineType(\'' + escapeHtml(item.type) + '\', this.checked)">',
-      '<span>' + escapeHtml(item.label) + '</span>',
-      '</label>',
-    ].join('');
-  }).join('');
-}
-
-function setTimelineRange(value) {
-  const minutes = parseInt(value, 10);
-  eventTimelineRange = [60, 240, 1440, 10080, 43200, 129600].includes(minutes) ? minutes : 60;
-  refreshEventTimeline();
-}
-
-function toggleTimelineType(type, checked) {
-  if (checked) {
-    if (!eventTimelineTypes.includes(type)) eventTimelineTypes.push(type);
-  } else {
-    eventTimelineTypes = eventTimelineTypes.filter(item => item !== type);
-  }
-  if (!eventTimelineTypes.length) eventTimelineTypes = EVENT_TIMELINE_TYPE_DEFS.map(item => item.type);
-  renderTimelineTypeFilters();
-  refreshEventTimeline();
-}
-
-function setTimelineStatus(message, type) {
-  const el = document.getElementById('timelineStatus');
-  if (!el) return;
-  el.textContent = message || '';
-  el.className = 'timeline-status' + (type ? ' ' + type : '');
-}
-
-function timelineDateFromValue(value) {
-  const raw = String(value || '').trim();
-  if (!raw) return null;
-  let date = new Date(raw);
-  if (!Number.isNaN(date.getTime())) return date;
-  date = new Date(raw.replace(' ', 'T'));
-  if (!Number.isNaN(date.getTime())) return date;
-  if (/^\d{1,2}:\d{2}/.test(raw)) {
-    const today = new Date();
-    const dateText = today.getFullYear() + '-' + String(today.getMonth() + 1).padStart(2, '0') + '-' + String(today.getDate()).padStart(2, '0');
-    date = new Date(dateText + 'T' + raw);
-    if (!Number.isNaN(date.getTime())) return date;
-  }
-  return null;
-}
-
-function timelineRangeForTimestamp(timestamp) {
-  const date = timelineDateFromValue(timestamp);
-  if (!date) return eventTimelineRange || 60;
-  const diffMinutes = Math.max(0, Math.ceil((Date.now() - date.getTime()) / 60000));
-  if (diffMinutes <= 60) return 60;
-  if (diffMinutes <= 240) return 240;
-  if (diffMinutes <= 1440) return 1440;
-  if (diffMinutes <= 10080) return 10080;
-  if (diffMinutes <= 43200) return 43200;
-  return 129600;
-}
-
-function timelineFocusTimeKey(value) {
-  const raw = String(value || '').trim();
-  if (!raw) return '';
-  if (/^\d{1,2}:\d{2}/.test(raw)) return raw.slice(0, 8);
-  return raw.replace('T', ' ').slice(0, 19);
-}
-
-function eventMatchesTimelineFocus(event, focus) {
-  if (!event || !focus) return false;
-  if (focus.type && event.type !== focus.type) return false;
-  const payload = event.payload && typeof event.payload === 'object' ? event.payload : {};
-  if (focus.sourceId && [event.id, payload.id, payload.note_id].some(value => String(value || '') === String(focus.sourceId))) return true;
-  const eventTime = timelineFocusTimeKey(event.timestamp);
-  const focusTime = timelineFocusTimeKey(focus.timestamp);
-  if (!eventTime || !focusTime) return false;
-  if (eventTime === focusTime) return true;
-  if (focusTime.length <= 8 && eventTime.slice(11, 19).startsWith(focusTime.slice(0, 5))) return true;
-  return false;
-}
-
-function openEventTimelineAround(timestamp, type, sourceId) {
-  pendingTimelineFocus = {
-    type: type || '',
-    sourceId: sourceId == null ? '' : String(sourceId),
-    timestamp: timestamp || '',
-  };
-  selectedTimelineEventId = null;
-  eventTimelineRange = timelineRangeForTimestamp(timestamp);
-  eventTimelineTypes = EVENT_TIMELINE_TYPE_DEFS.map(item => item.type);
-  const rangeSelect = document.getElementById('timelineRange');
-  if (rangeSelect) rangeSelect.value = String(eventTimelineRange);
-  document.getElementById('historyBackdrop').classList.add('show');
-  renderTimelineTypeFilters();
-  switchHistoryView('timeline', true);
-  refreshHistory();
-  refreshEventTimeline();
-}
-
-function refreshEventTimeline() {
-  setTimelineStatus('正在加载事件时间轴...', '');
-  socket.emit('get_event_timeline', {
-    minutes: eventTimelineRange,
-    limit: 300,
-    types: eventTimelineTypes,
-  });
-}
-
-function applyEventTimeline(data) {
-  eventTimelineState = Object.assign({ events: [], summary: {}, filters: {}, range: {}, price_summary: {} }, data || {});
-  const events = Array.isArray(eventTimelineState.events) ? eventTimelineState.events : [];
-  let focusMissing = false;
-  if (pendingTimelineFocus) {
-    const focused = events.find(event => eventMatchesTimelineFocus(event, pendingTimelineFocus));
-    if (focused) selectedTimelineEventId = focused.id;
-    else focusMissing = true;
-    pendingTimelineFocus = null;
-  } else if (selectedTimelineEventId && !events.some(event => event.id === selectedTimelineEventId)) {
-    selectedTimelineEventId = null;
-  }
-  setTimelineStatus(focusMissing ? '已打开复盘时间轴，未在当前范围找到对应事件。' : '', focusMissing ? 'fail' : '');
-  renderEventTimeline();
-  if (selectedTimelineEventId) {
-    requestAnimationFrame(() => {
-      const active = document.querySelector('#timelineList .timeline-event.active');
-      if (active) active.scrollIntoView({ block: 'center', behavior: 'smooth' });
-    });
-  }
-}
-
-function timelineEventTime(timestamp) {
-  const text = String(timestamp || '');
-  if (!text) return '--';
-  return text.replace('T', ' ').slice(0, 19);
-}
-
-function renderTimelineSummary() {
-  const box = document.getElementById('timelineSummary');
-  if (!box) return;
-  const summary = eventTimelineState.summary || {};
-  const byType = summary.by_type || {};
-  const range = eventTimelineState.range || {};
-  const statItems = [
-    ['事件', summary.total || 0],
-    ['预警', byType.alert || 0],
-    ['笔记', byType.review_note || 0],
-    ['范围', range.minutes ? Math.round(Number(range.minutes) / 60) + 'h' : '--'],
-  ];
-  box.innerHTML = statItems.map(item => (
-    '<div class="history-stat"><div class="history-stat-label">' + escapeHtml(item[0]) + '</div><div class="history-stat-value">' + escapeHtml(String(item[1])) + '</div></div>'
-  )).join('');
-}
-
-function renderTimelineList() {
-  const list = document.getElementById('timelineList');
-  if (!list) return;
-  const events = Array.isArray(eventTimelineState.events) ? eventTimelineState.events : [];
-  if (!events.length) {
-    list.innerHTML = '<div class="history-empty">暂无事件</div>';
-    return;
-  }
-  list.innerHTML = events.slice().reverse().map(event => [
-    '<button class="timeline-event' + (selectedTimelineEventId === event.id ? ' active' : '') + '" type="button" onclick="selectTimelineEvent(decodeURIComponent(\'' + encodeURIComponent(String(event.id || '')) + '\'))">',
-    '<span class="timeline-event-time">' + escapeHtml(timelineEventTime(event.timestamp).slice(11) || '--') + '</span>',
-    '<span class="timeline-event-main">',
-    '<span class="timeline-event-title">' + escapeHtml(event.title || timelineTypeLabel(event.type)) + '</span>',
-    String(event.summary || '').trim() && String(event.summary || '').trim() !== String(event.title || '').trim()
-      ? '<span class="timeline-event-summary">' + escapeHtml(event.summary) + '</span>'
-      : '',
-    '<span class="timeline-event-type">' + escapeHtml(timelineTypeLabel(event.type)) + '</span>',
-    '</span>',
-    '</button>',
-  ].join('')).join('');
-}
-
-function selectTimelineEvent(id) {
-  selectedTimelineEventId = id;
-  renderTimelineList();
-  renderTimelineDetail();
-}
-
-function detailCell(label, value) {
-  const text = value == null || value === '' ? '暂无详情' : String(value);
-  return '<div class="timeline-detail-cell"><span>' + escapeHtml(label) + '</span><strong>' + escapeHtml(text) + '</strong></div>';
-}
-
-function renderTimelineDetail() {
-  const detail = document.getElementById('timelineDetail');
-  if (!detail) return;
-  const events = Array.isArray(eventTimelineState.events) ? eventTimelineState.events : [];
-  const event = events.find(item => item.id === selectedTimelineEventId);
-  if (!event) {
-    detail.innerHTML = '<div class="history-empty">选择一条事件查看详情</div>';
-    return;
-  }
-  const payload = event.payload && typeof event.payload === 'object' ? event.payload : {};
-  const cells = [
-    detailCell('类型', timelineTypeLabel(event.type)),
-    detailCell('来源', event.source),
-    detailCell('时间', timelineEventTime(event.timestamp)),
-  ];
-  let extras = '';
-  if (event.type === 'alert') {
-    cells.push(detailCell('等级', alertLevelLabel(payload.level)));
-    cells.push(detailCell('品种', alertModeLabel(payload.mode)));
-    cells.push(detailCell('处置结果', payload.handled ? '已处理' : '未处理'));
-    if (payload.handled_at) cells.push(detailCell('处理时间', payload.handled_at));
-    if (payload.handling_note) cells.push(detailCell('处理备注', payload.handling_note));
-    if (Array.isArray(payload.related_news) && payload.related_news.length) {
-      extras += '<div class="timeline-detail-news">' + payload.related_news.slice(0, 3).map(item => (
-        '<a href="' + escapeHtml(item.url || '#') + '" target="_blank" rel="noopener noreferrer">' + escapeHtml(item.title || '相关新闻') + '</a>'
-      )).join('') + '</div>';
-    }
-  } else if (event.type === 'risk_analysis') {
-    const structured = payload.structured || {};
-    const quality = payload.market_quality || payload.data_quality || {};
-    cells.push(detailCell('模型', [payload.provider, payload.model].filter(Boolean).join(' / ')));
-    cells.push(detailCell('行情质量', quality.score == null ? '' : quality.score + '分'));
-    cells.push(detailCell('风险等级', structured.risk_level || ''));
-    cells.push(detailCell('主要因素', structured.key_factors || structured.main_factors || ''));
-  } else if (event.type === 'news') {
-    cells.push(detailCell('来源', payload.source));
-    cells.push(detailCell('主题', payload.topic));
-    if (payload.url) {
-      extras += '<div class="timeline-detail-news"><a href="' + escapeHtml(payload.url) + '" target="_blank" rel="noopener noreferrer">' + escapeHtml(payload.url) + '</a></div>';
-    }
-  } else if (event.type === 'data_status') {
-    cells.push(detailCell('状态', payload.cached ? '缓存' : (payload.ok === false ? '异常' : payload.status)));
-    cells.push(detailCell('说明', payload.error || payload.message));
-    if (payload.summary) cells.push(detailCell('价差比例', payload.summary.spread_pct == null ? '' : payload.summary.spread_pct + '%'));
-  } else if (event.type === 'price_summary') {
-    const summary = payload.summary || {};
-    const usd = summary.usd || {};
-    const rmb = summary.rmb || {};
-    cells.push(detailCell('价格点', payload.points));
-    cells.push(detailCell('USD/oz', usd.start == null ? '' : usd.start + ' -> ' + usd.end));
-    cells.push(detailCell('RMB/克', rmb.start == null ? '' : rmb.start + ' -> ' + rmb.end));
-  } else if (event.type === 'review_note') {
-    if (payload.updated_at) cells.push(detailCell('最后更新', timelineEventTime(payload.updated_at)));
-    if (payload.related_event_id) {
-      cells.push(detailCell('关联类型', timelineTypeLabel(payload.related_event_type)));
-      cells.push(detailCell('关联事件', payload.related_event_title || payload.related_event_id));
-    }
-  }
-  const actions = event.type === 'review_note'
-    ? [
-      '<div class="timeline-detail-actions">',
-      '<button class="settings-cancel" type="button" onclick="editSelectedReviewNote()">编辑笔记</button>',
-      '<button class="dialog-danger" type="button" onclick="deleteSelectedReviewNote()">删除笔记</button>',
-      '</div>',
-    ].join('')
-    : '<div class="timeline-detail-actions"><button class="settings-cancel" type="button" onclick="openReviewNoteEditorFromSelectedEvent()">关联创建笔记</button></div>';
-  detail.innerHTML = [
-    '<div class="timeline-detail-title">' + escapeHtml(event.title || timelineTypeLabel(event.type)) + '</div>',
-    '<div class="timeline-detail-meta">' + escapeHtml(timelineEventTime(event.timestamp)) + ' · ' + escapeHtml(event.source || '--') + '</div>',
-    '<div class="timeline-detail-summary">' + escapeHtml(payload.message || payload.content || event.summary || '暂无详情') + '</div>',
-    '<div class="timeline-detail-grid">' + cells.join('') + '</div>',
-    extras,
-    actions,
-  ].join('');
-}
-
-function renderEventTimeline() {
-  renderTimelineSummary();
-  renderTimelineList();
-  renderTimelineDetail();
-}
-
-function exportReviewReport() {
-  setTimelineStatus('正在导出复盘报告...', '');
-  socket.emit('export_review_report', {
-    minutes: eventTimelineRange,
-    limit: 300,
-    types: eventTimelineTypes,
-  });
-}
-
-function exportHistoryCsv() {
-  socket.emit('export_price_history', {});
-}
 
 function setOpsStatus(message, ok) {
   const el = document.getElementById('opsStatus');
@@ -3485,175 +2743,7 @@ function testAlert() {
   socket.emit('test_alert', { type: 'warning' });
 }
 
-// ========== 风险分析助手 ==========
-function selectedRiskPrice(data) {
-  const source = data || latestData || {};
-  return currentMode === 'usd' ? source.usd : source.rmb;
-}
-
-function hasRiskAnalysisInput(data) {
-  const price = Number(selectedRiskPrice(data));
-  return Number.isFinite(price) && price > 0;
-}
-
-function riskAnalysisUnavailableMessage() {
-  if (!hasRiskAnalysisInput()) return '当前没有可用于风险分析的行情价格，请先重新获取行情数据。';
-  const retry = document.getElementById('priceRetry');
-  if (retry && retry.classList.contains('show')) return '当前行情状态异常，请先重新获取行情数据后再分析。';
-  return '';
-}
-
-function updateRiskEntryState() {
-  const riskAnalyzeButton = document.getElementById('riskAnalyzeButton');
-  const available = !riskAnalysisUnavailableMessage();
-  if (riskAnalyzeButton) {
-    riskAnalyzeButton.hidden = !available;
-    riskAnalyzeButton.disabled = !available || !appSettings.risk_assistant_enabled;
-    riskAnalyzeButton.setAttribute('aria-hidden', available ? 'false' : 'true');
-  }
-  document.querySelectorAll('.source-risk-action').forEach(button => {
-    button.hidden = !available;
-    button.disabled = !available || !appSettings.risk_assistant_enabled;
-    button.setAttribute('aria-hidden', available ? 'false' : 'true');
-  });
-}
-
-function openRiskAnalysis() {
-  const riskUnavailable = riskAnalysisUnavailableMessage();
-  if (riskUnavailable) {
-    applyFetchStatus({ ok:false, message:riskUnavailable, retryable:true });
-    return false;
-  }
-  document.getElementById('riskBackdrop').classList.add('show');
-  socket.emit('get_risk_analysis_history');
-  const providerMessage = riskProviderErrorMessage();
-  if (!appSettings.risk_assistant_enabled) {
-    applyRiskStatus('风险分析助手已关闭，请先在设置中启用。', 'error');
-  } else if (providerMessage) {
-    applyRiskStatus(providerMessage, 'error');
-  } else if (!document.getElementById('riskResult').textContent || document.getElementById('riskResult').textContent === '暂无分析结果。') {
-    applyRiskStatus('点击开始分析，助手会基于当前行情生成风险趋势判断。', '');
-  }
-  updateRiskButtonState();
-  return true;
-}
-
-function closeRiskAnalysis() {
-  document.getElementById('riskBackdrop').classList.remove('show');
-}
-
-function onRiskBackdrop(event) {
-  if (event.target.id === 'riskBackdrop') closeRiskAnalysis();
-}
-
-function applyRiskStatus(message, type) {
-  const statusEl = document.getElementById('riskStatus');
-  statusEl.textContent = message || '';
-  statusEl.className = 'risk-status' + (type ? ' ' + type : '');
-}
-
-function renderRiskDiagnostic(diagnostic) {
-  const el = document.getElementById('riskDiagnostic');
-  if (!el) return;
-  if (!diagnostic) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  const provider = [diagnostic.provider, diagnostic.model].filter(Boolean).join(' / ');
-  const recovery = Array.isArray(diagnostic.recovery) ? diagnostic.recovery.filter(Boolean) : [];
-  const recoveryHtml = recovery.length
-    ? '<ul class="risk-diagnostic-list">' + recovery.map(item => '<li>' + escapeHtml(item) + '</li>').join('') + '</ul>'
-    : '<div>' + escapeHtml('稍后重试；如果问题持续，请导出诊断报告核对配置和网络状态。') + '</div>';
-  el.innerHTML = [
-    '<div class="risk-diagnostic-title">失败原因 · ' + escapeHtml(diagnostic.title || '风险分析失败') + '</div>',
-    provider ? '<div class="risk-diagnostic-meta">模型 ' + escapeHtml(provider) + '</div>' : '',
-    '<div class="risk-diagnostic-section"><strong>原因</strong><div>' + escapeHtml(diagnostic.reason || '未知错误') + '</div></div>',
-    '<div class="risk-diagnostic-section"><strong>影响</strong><div>' + escapeHtml(diagnostic.impact || '风险分析未生成，本次不会写入分析历史。') + '</div></div>',
-    '<div class="risk-diagnostic-section"><strong>建议处理</strong>' + recoveryHtml + '</div>',
-  ].join('');
-  el.classList.add('show');
-}
-
-function updateRiskButtonState() {
-  const runBtn = document.getElementById('riskRunButton');
-  const forceBtn = document.getElementById('riskForceRunButton');
-  const riskUnavailable = riskAnalysisUnavailableMessage();
-  const disabled = riskAnalysisRunning || !appSettings.risk_assistant_enabled || !!riskProviderErrorMessage() || !!riskUnavailable;
-  if (runBtn) {
-    runBtn.disabled = disabled;
-    runBtn.textContent = riskAnalysisRunning ? '分析中...' : '开始分析';
-  }
-  if (forceBtn) forceBtn.disabled = disabled;
-  updateRiskEntryState();
-}
-
-function requestRiskAnalysis(trigger, force) {
-  if (riskAnalysisRunning) return;
-  const riskUnavailable = riskAnalysisUnavailableMessage();
-  if (riskUnavailable) {
-    applyRiskStatus(riskUnavailable, 'error');
-    updateRiskButtonState();
-    return;
-  }
-  if (!appSettings.risk_assistant_enabled) {
-    applyRiskStatus('风险分析助手已关闭，请先在设置中启用。', 'error');
-    return;
-  }
-  const providerMessage = riskProviderErrorMessage();
-  if (providerMessage) {
-    applyRiskStatus(providerMessage, 'error');
-    return;
-  }
-  riskAnalysisRunning = true;
-  document.getElementById('riskResult').textContent = '正在分析当前行情...';
-  renderRiskDiagnostic(null);
-  renderRiskSnapshot(null);
-  renderRiskStructured(null);
-  pendingRiskForceTrigger = null;
-  setRiskForceButtonVisible(false);
-  applyRiskStatus('正在生成风险分析...', 'loading');
-  updateRiskButtonState();
-  socket.emit('request_risk_analysis', Object.assign({}, trigger ? { trigger } : {}, force ? { force: true } : {}));
-}
-
-function setRiskForceButtonVisible(visible) {
-  const btn = document.getElementById('riskForceRunButton');
-  if (btn) btn.style.display = visible ? '' : 'none';
-}
-
-function rerunRiskAnalysis() {
-  requestRiskAnalysis(pendingRiskForceTrigger || null, true);
-}
-
-function riskProviderErrorMessage() {
-  const provider = appSettings.risk_assistant_provider || 'deepseek';
-  if (provider === 'deepseek') {
-    if (!appSettings.deepseek_api_key_configured) return '请先在设置中配置 DeepSeek API Key。';
-    if (!appSettings.deepseek_model) return '请先选择 DeepSeek 模型。';
-    return '';
-  }
-  if (provider === 'openai_compatible') {
-    if (!appSettings.openai_compatible_base_url) return '请先配置兼容接口地址。';
-    if (!appSettings.openai_compatible_model) return '请先配置兼容模型。';
-    if (!appSettings.openai_compatible_api_key_configured) return '请先配置兼容 API Key。';
-    return '';
-  }
-  return '当前模型提供商暂不支持。';
-}
-
-function formatRiskNumber(value, suffix) {
-  if (value === null || value === undefined || value === '') return '--';
-  const num = Number(value);
-  if (!Number.isFinite(num)) return String(value);
-  return num.toLocaleString('en-US', { maximumFractionDigits: 2 }) + (suffix || '');
-}
-
-function formatRiskUsage(usage) {
-  if (!usage || usage.total_tokens == null) return '';
-  return '本次用量 ' + usage.total_tokens + ' tokens。';
-}
-
+// ========== 通用界面工具 ==========
 function escapeHtml(value) {
   return String(value || '').replace(/[&<>"']/g, ch => ({
     '&': '&amp;',
@@ -3693,464 +2783,6 @@ function setupEllipsisTooltips() {
 }
 
 setupEllipsisTooltips();
-
-function renderRiskEvidence(snapshot) {
-  const el = document.getElementById('riskEvidence');
-  if (!el) return;
-  if (!snapshot) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  const evidence = snapshot.evidence_summary || {};
-  const goldCached = evidence.gold_cached ? '缓存' : '实时';
-  const rateCached = evidence.rate_cached ? '缓存' : '实时';
-  const qualityLabel = evidence.quality_label || '--';
-  const qualityScore = evidence.quality_score == null ? '--' : evidence.quality_score + '分';
-  const items = [
-    ['当前金价', 'RMB ' + formatRiskNumber(evidence.price_rmb ?? snapshot.price_rmb, '/克') + ' · USD ' + formatRiskNumber(evidence.price_usd ?? snapshot.price_usd, '/oz')],
-    ['行情源', (evidence.gold_source || snapshot.gold_source || '--') + ' · ' + goldCached + (evidence.gold_time ? ' · ' + evidence.gold_time : '')],
-    ['汇率源', (evidence.rate_source || snapshot.rate_source || '--') + ' · ' + rateCached + (evidence.rate_time ? ' · ' + evidence.rate_time : '')],
-    ['样本规模', '历史 ' + (evidence.history_points ?? snapshot.history_points ?? 0) + ' 点 · 5分钟K线 ' + (evidence.kline_points ?? snapshot.kline_points ?? 0) + ' 根'],
-    ['近期资讯', (evidence.news_count ?? snapshot.news_count ?? 0) + ' 条'],
-    ['数据质量', qualityScore + ' · ' + qualityLabel],
-  ];
-  const missing = Array.isArray(evidence.missing) ? evidence.missing.filter(Boolean) : [];
-  const recovery = Array.isArray(evidence.recovery) ? evidence.recovery.filter(Boolean) : [];
-  const warning = missing.length || recovery.length
-    ? '<div class="risk-evidence-warning"><strong>缺失数据</strong>：' + escapeHtml(missing.join('、') || '暂无') + '<br><strong>恢复建议</strong>：' + escapeHtml(recovery.join('；') || '继续等待数据更新') + '</div>'
-    : '';
-  const qualitySummary = evidence.quality_summary ? '<div class="risk-evidence-warning">' + escapeHtml(evidence.quality_summary) + '</div>' : '';
-  el.innerHTML = [
-    '<div class="risk-block-title">数据依据</div>',
-    '<div class="risk-evidence-grid">',
-    items.map(item => '<div class="risk-evidence-item"><div class="risk-evidence-label">' + escapeHtml(item[0]) + '</div><div class="risk-evidence-value">' + escapeHtml(item[1]) + '</div></div>').join(''),
-    '</div>',
-    warning + qualitySummary,
-  ].join('');
-  el.classList.add('show');
-}
-
-function renderRiskSnapshot(snapshot) {
-  const meta = document.getElementById('riskMeta');
-  if (!snapshot) {
-    meta.innerHTML = '';
-    meta.classList.remove('show');
-    renderRiskEvidence(null);
-    renderRiskQuality(null);
-    renderRiskTrends(null);
-    renderRiskScorecard(null);
-    renderRiskStructured(null);
-    return;
-  }
-  const items = [
-    '时间 ' + (snapshot.analysis_time || '--'),
-    'RMB ' + formatRiskNumber(snapshot.price_rmb, '/克'),
-    'USD ' + formatRiskNumber(snapshot.price_usd, '/oz'),
-    '汇率 ' + formatRiskNumber(snapshot.usdcny_rate, ''),
-    '历史点 ' + (snapshot.history_points || 0),
-    'K线 ' + (snapshot.kline_points || 0),
-    '资讯 ' + (snapshot.news_count || 0),
-  ];
-  const marketQuality = snapshot.market_quality && Object.keys(snapshot.market_quality).length ? snapshot.market_quality : null;
-  const quality = marketQuality || snapshot.data_quality || null;
-  if (quality) items.push('行情质量 ' + quality.score + '分/' + (quality.label || quality.level || '--'));
-  if (snapshot.sample_warning) items.push(snapshot.sample_warning);
-  meta.innerHTML = items.map(item => '<span>' + escapeHtml(item) + '</span>').join('');
-  meta.classList.add('show');
-  renderRiskEvidence(snapshot);
-  renderRiskQuality(quality);
-  renderRiskTrends(snapshot.multi_period_trends || []);
-  renderRiskScorecard(snapshot.risk_scorecard || null);
-}
-
-function renderRiskQuality(quality) {
-  const el = document.getElementById('riskQuality');
-  if (!quality) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  const reasons = Array.isArray(quality.reasons) ? quality.reasons.filter(Boolean).join('；') : '';
-  const summary = quality.summary || reasons || quality.label || '暂无说明';
-  el.innerHTML = [
-    '<div class="risk-block-title">行情质量</div>',
-    '<div class="risk-quality-score">' + escapeHtml(quality.score == null ? '--' : quality.score) + '</div>',
-    '<div class="risk-quality-level">等级 ' + escapeHtml(quality.level || '--') + '</div>',
-    '<div class="risk-quality-summary">' + escapeHtml(summary) + '</div>',
-  ].join('');
-  el.classList.add('show');
-}
-
-function riskTrendClass(direction) {
-  if (direction === '上行') return 'up';
-  if (direction === '下行') return 'down';
-  if (direction === '震荡') return 'flat';
-  return 'missing';
-}
-
-function renderRiskTrends(trends) {
-  const el = document.getElementById('riskTrends');
-  if (!Array.isArray(trends) || !trends.length) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  const items = trends.map(item => {
-    const direction = item.direction_rmb || item.direction_usd || '样本不足';
-    const pct = item.rmb && item.rmb.change_pct != null ? item.rmb.change_pct : item.usd && item.usd.change_pct;
-    return [
-      '<div class="risk-trend-item">',
-      '<div class="risk-trend-period">' + escapeHtml(item.minutes || '--') + '分钟 · ' + escapeHtml(item.points || 0) + '点</div>',
-      '<div class="risk-trend-direction ' + riskTrendClass(direction) + '">' + escapeHtml(direction) + '</div>',
-      '<div class="risk-trend-change">变动 ' + escapeHtml(pct == null ? '--' : Number(pct).toFixed(2) + '%') + '</div>',
-      '</div>',
-    ].join('');
-  }).join('');
-  el.innerHTML = '<div class="risk-block-title">多周期趋势</div><div class="risk-trend-list">' + items + '</div>';
-  el.classList.add('show');
-}
-
-function renderRiskScorecard(scorecard) {
-  const el = document.getElementById('riskScorecard');
-  if (!scorecard) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  const items = [
-    ['总体风险', scorecard.overall_risk],
-    ['趋势强度', scorecard.trend_strength],
-    ['波动风险', scorecard.volatility_risk],
-    ['汇率影响', scorecard.fx_impact],
-    ['事件风险', scorecard.event_risk],
-    ['数据可信度', scorecard.data_credibility],
-  ];
-  el.innerHTML = '<div class="risk-block-title">风险评分卡</div><div class="risk-score-grid">' + items.map(item => (
-    '<div class="risk-score-item"><div class="risk-score-label">' + escapeHtml(item[0]) + '</div><div class="risk-score-value">' + escapeHtml(item[1] == null ? '--' : item[1]) + '</div></div>'
-  )).join('') + '</div>';
-  el.classList.add('show');
-}
-
-function renderRiskStructured(structured) {
-  const el = document.getElementById('riskStructured');
-  if (!el) return;
-  const labels = [
-    ['risk_level', '风险等级'],
-    ['trend_direction', '趋势方向'],
-    ['data_credibility', '数据可信度'],
-    ['main_factors', '主要影响因素'],
-    ['watch_range', '观察价格区间'],
-    ['follow_up', '后续关注'],
-  ];
-  const items = labels
-    .map(([key, label]) => [label, structured && structured[key]])
-    .filter(([, value]) => value);
-  if (!items.length) {
-    el.innerHTML = '';
-    el.classList.remove('show');
-    return;
-  }
-  el.innerHTML = items.map(([label, value]) => (
-    '<div class="risk-structured-item"><div class="risk-structured-label">' + escapeHtml(label) + '</div><div class="risk-structured-value">' + escapeHtml(value) + '</div></div>'
-  )).join('');
-  el.classList.add('show');
-}
-
-function applyRiskHistory(data) {
-  riskAnalysisHistory = Array.isArray(data && data.items) ? data.items : [];
-  riskComparisonSelection = [];
-  const comparison = document.getElementById('riskComparison');
-  if (comparison) {
-    comparison.innerHTML = '';
-    comparison.hidden = true;
-  }
-  renderRiskHistory();
-}
-
-function renderRiskHistory() {
-  const list = document.getElementById('riskHistoryList');
-  const clearBtn = document.getElementById('riskClearHistoryButton');
-  const compareBtn = document.getElementById('riskCompareButton');
-  if (clearBtn) clearBtn.disabled = riskAnalysisHistory.length === 0;
-  if (compareBtn) {
-    compareBtn.disabled = riskComparisonSelection.length !== 2;
-    compareBtn.textContent = riskComparisonSelection.length
-      ? '对比所选（' + riskComparisonSelection.length + '/2）'
-      : '对比所选';
-  }
-  if (!riskAnalysisHistory.length) {
-    list.innerHTML = '<div class="risk-history-empty">暂无历史记录</div>';
-    return;
-  }
-  list.innerHTML = riskAnalysisHistory.map((item, index) => {
-    const firstLine = String(item.content || '').split('\n').find(Boolean) || '历史分析';
-    const qualitySource = item.snapshot ? (item.snapshot.market_quality || item.snapshot.data_quality) : null;
-    const quality = qualitySource ? ' · 行情质量 ' + qualitySource.score + '分' : '';
-    const evidence = item.evidence_summary || (item.snapshot && item.snapshot.evidence_summary) || {};
-    const samples = evidence.history_points != null || evidence.kline_points != null
-      ? ' · 历史' + (evidence.history_points ?? 0) + '/K线' + (evidence.kline_points ?? 0)
-      : '';
-    const selectedForComparison = riskComparisonSelection.includes(index);
-    return [
-      '<div class="risk-history-item' + (selectedForComparison ? ' selected' : '') + '">',
-      '<button class="risk-history-main" type="button" onclick="openRiskHistoryItem(' + index + ')">',
-      '<div class="risk-history-time">' + escapeHtml(item.analysis_time || '--') + escapeHtml(quality) + escapeHtml(samples) + '</div>',
-      '<div class="risk-history-text">' + escapeHtml(firstLine) + '</div>',
-      '</button>',
-      '<button class="btn-clear-sm btn-muted-sm risk-history-compare" type="button" aria-pressed="' + String(selectedForComparison) + '" onclick="toggleRiskComparisonItem(' + index + ')">' + (selectedForComparison ? '已选' : '选择对比') + '</button>',
-      '<button class="btn-clear-sm btn-muted-sm risk-history-review" type="button" data-risk-timeline-index="' + index + '" onclick="window.openRiskTimelineFromHistory(' + index + ')">查看复盘</button>',
-      '</div>',
-    ].join('');
-  }).join('');
-}
-
-function clearRenderedRiskComparison() {
-  const comparison = document.getElementById('riskComparison');
-  if (!comparison) return;
-  comparison.innerHTML = '';
-  comparison.hidden = true;
-}
-
-function toggleRiskComparisonItem(index) {
-  if (!Number.isInteger(index) || !riskAnalysisHistory[index]) return;
-  const selectedIndex = riskComparisonSelection.indexOf(index);
-  if (selectedIndex >= 0) {
-    riskComparisonSelection.splice(selectedIndex, 1);
-    clearRenderedRiskComparison();
-    applyRiskStatus('已取消该条对比选择。', '');
-    renderRiskHistory();
-    return;
-  }
-  if (riskComparisonSelection.length >= 2) {
-    applyRiskStatus('最多选择两条风险分析，请先取消一条已选记录。', 'error');
-    return;
-  }
-  riskComparisonSelection.push(index);
-  clearRenderedRiskComparison();
-  applyRiskStatus(
-    riskComparisonSelection.length === 2 ? '已选择两条风险分析，可以开始对比。' : '已选择一条风险分析，请再选择一条。',
-    ''
-  );
-  renderRiskHistory();
-}
-
-function riskComparisonValue(value) {
-  if (value === null || value === undefined || value === '') return '暂无';
-  if (typeof value === 'number' && !Number.isFinite(value)) return '暂无';
-  return String(value);
-}
-
-function riskComparisonNumber(value, suffix) {
-  if (value === null || value === undefined || value === '') return '暂无';
-  const number = Number(value);
-  if (!Number.isFinite(number)) return '暂无';
-  return number.toLocaleString('en-US', { maximumFractionDigits: 4 }) + (suffix || '');
-}
-
-function riskComparisonSource(snapshot, sourceKey, cachedKey) {
-  const source = snapshot && snapshot[sourceKey];
-  if (!source) return '暂无';
-  const cached = snapshot[cachedKey];
-  const state = cached === true ? '缓存' : cached === false ? '实时' : '状态暂无';
-  return String(source) + ' · ' + state;
-}
-
-function riskComparisonQuality(snapshot) {
-  if (!snapshot || typeof snapshot !== 'object') return '暂无';
-  const marketQuality = snapshot.market_quality && Object.keys(snapshot.market_quality).length
-    ? snapshot.market_quality
-    : null;
-  const quality = marketQuality || (snapshot.data_quality && Object.keys(snapshot.data_quality).length ? snapshot.data_quality : null);
-  if (!quality) return '暂无';
-  const score = riskComparisonValue(quality.score);
-  const label = riskComparisonValue(quality.label || quality.level);
-  if (score === '暂无' && label === '暂无') return '暂无';
-  if (score === '暂无') return label;
-  if (label === '暂无') return score + '分';
-  return score + '分 · ' + label;
-}
-
-function riskComparisonSamples(snapshot) {
-  if (!snapshot || typeof snapshot !== 'object') return '暂无';
-  const values = [
-    ['历史', snapshot.history_points, '点'],
-    ['K线', snapshot.kline_points, '根'],
-    ['资讯', snapshot.news_count, '条'],
-  ].filter(item => item[1] !== null && item[1] !== undefined && item[1] !== '');
-  if (!values.length) return '暂无';
-  return values.map(item => item[0] + ' ' + riskComparisonNumber(item[1], item[2])).join(' · ');
-}
-
-function riskComparisonEntryTime(item) {
-  if (!item || typeof item !== 'object') return '';
-  return item.analysis_time || (item.snapshot && item.snapshot.analysis_time) || '';
-}
-
-function riskComparisonParsedTime(item) {
-  const value = riskComparisonEntryTime(item);
-  if (!value) return null;
-  const parsed = Date.parse(value);
-  return Number.isFinite(parsed) ? parsed : null;
-}
-
-function riskComparisonRow(field, label, earlierValue, laterValue, sideLabels) {
-  const earlierText = riskComparisonValue(earlierValue);
-  const laterText = riskComparisonValue(laterValue);
-  const changedClass = earlierText === laterText ? '' : ' changed';
-  const labels = sideLabels || { earlier: '较早', later: '较新' };
-  return [
-    '<div class="risk-comparison-row' + changedClass + '" data-field="' + escapeHtml(field) + '">',
-    '<div class="risk-comparison-label">' + escapeHtml(label) + '</div>',
-    '<div class="risk-comparison-value" data-side="earlier" data-side-label="' + escapeHtml(labels.earlier) + '">' + escapeHtml(earlierText) + '</div>',
-    '<div class="risk-comparison-value" data-side="later" data-side-label="' + escapeHtml(labels.later) + '">' + escapeHtml(laterText) + '</div>',
-    '</div>',
-  ].join('');
-}
-
-function renderRiskComparison(earlier, later, sideLabels) {
-  const comparison = document.getElementById('riskComparison');
-  if (!comparison) return;
-  const labels = sideLabels || { earlier: '较早', later: '较新' };
-  const earlierSnapshot = earlier && earlier.snapshot && typeof earlier.snapshot === 'object' ? earlier.snapshot : {};
-  const laterSnapshot = later && later.snapshot && typeof later.snapshot === 'object' ? later.snapshot : {};
-  const earlierStructured = earlier && earlier.structured && typeof earlier.structured === 'object' ? earlier.structured : {};
-  const laterStructured = later && later.structured && typeof later.structured === 'object' ? later.structured : {};
-  const earlierScorecard = earlierSnapshot.risk_scorecard && typeof earlierSnapshot.risk_scorecard === 'object' ? earlierSnapshot.risk_scorecard : {};
-  const laterScorecard = laterSnapshot.risk_scorecard && typeof laterSnapshot.risk_scorecard === 'object' ? laterSnapshot.risk_scorecard : {};
-  const structuredFields = [
-    ['risk_level', '风险等级'],
-    ['trend_direction', '趋势方向'],
-    ['data_credibility', '数据可信度'],
-    ['main_factors', '主要影响因素'],
-    ['watch_range', '观察价格区间'],
-    ['follow_up', '后续关注'],
-  ];
-  const scorecardFields = [
-    ['overall_risk', '总体风险'],
-    ['trend_strength', '趋势强度'],
-    ['volatility_risk', '波动风险'],
-    ['fx_impact', '汇率影响'],
-    ['event_risk', '事件风险'],
-    ['data_credibility', '数据可信度'],
-  ];
-  const snapshotFields = [
-    ['price_rmb', '人民币克价', item => riskComparisonNumber(item.price_rmb, ' 元/克')],
-    ['price_usd', '国际金价', item => riskComparisonNumber(item.price_usd, ' 美元/盎司')],
-    ['usdcny_rate', '美元人民币汇率', item => riskComparisonNumber(item.usdcny_rate, '')],
-    ['gold_source', '金价来源', item => riskComparisonSource(item, 'gold_source', 'gold_cached')],
-    ['rate_source', '汇率来源', item => riskComparisonSource(item, 'rate_source', 'rate_cached')],
-    ['quality', '行情质量', item => riskComparisonQuality(item)],
-    ['samples', '样本规模', item => riskComparisonSamples(item)],
-  ];
-  const section = (title, rows) => [
-    '<section class="risk-comparison-section">',
-    '<div class="risk-comparison-title">' + escapeHtml(title) + '</div>',
-    '<div class="risk-comparison-grid">' + rows.join('') + '</div>',
-    '</section>',
-  ].join('');
-  const earlierProvider = [earlier.provider, earlier.model].filter(Boolean).join(' / ') || '暂无';
-  const laterProvider = [later.provider, later.model].filter(Boolean).join(' / ') || '暂无';
-  comparison.innerHTML = [
-    '<div class="risk-comparison-head">',
-    '<div class="risk-block-title">风险分析对比</div>',
-    '<div class="risk-comparison-sides">',
-    '<div class="risk-comparison-side" data-side="earlier"><strong>' + escapeHtml(labels.earlier) + '</strong><span>' + escapeHtml(riskComparisonValue(riskComparisonEntryTime(earlier))) + '</span><span>' + escapeHtml(earlierProvider) + '</span></div>',
-    '<div class="risk-comparison-side" data-side="later"><strong>' + escapeHtml(labels.later) + '</strong><span>' + escapeHtml(riskComparisonValue(riskComparisonEntryTime(later))) + '</span><span>' + escapeHtml(laterProvider) + '</span></div>',
-    '</div>',
-    '</div>',
-    section('结构化字段', structuredFields.map(item => riskComparisonRow(item[0], item[1], earlierStructured[item[0]], laterStructured[item[0]], labels))),
-    section('风险评分卡', scorecardFields.map(item => riskComparisonRow(item[0], item[1], earlierScorecard[item[0]], laterScorecard[item[0]], labels))),
-    section('行情快照', snapshotFields.map(item => riskComparisonRow(item[0], item[1], item[2](earlierSnapshot), item[2](laterSnapshot), labels))),
-  ].join('');
-  comparison.hidden = false;
-}
-
-function compareSelectedRiskHistory() {
-  if (riskComparisonSelection.length !== 2) {
-    applyRiskStatus('请选择两条风险分析后再对比。', 'error');
-    return;
-  }
-  const selected = riskComparisonSelection.map(index => riskAnalysisHistory[index]);
-  if (selected.some(item => !item)) {
-    riskComparisonSelection = [];
-    clearRenderedRiskComparison();
-    renderRiskHistory();
-    applyRiskStatus('所选历史记录已更新，请重新选择两条风险分析。', 'error');
-    return;
-  }
-  let earlier = selected[0];
-  let later = selected[1];
-  const earlierTime = riskComparisonParsedTime(earlier);
-  const laterTime = riskComparisonParsedTime(later);
-  const timesAreComparable = earlierTime !== null && laterTime !== null;
-  const sideLabels = timesAreComparable
-    ? { earlier: '较早', later: '较新' }
-    : { earlier: '记录一', later: '记录二' };
-  if (timesAreComparable && earlierTime > laterTime) {
-    earlier = selected[1];
-    later = selected[0];
-  }
-  renderRiskComparison(earlier, later, sideLabels);
-  applyRiskStatus('已生成本地风险分析对比，不会调用模型。', '');
-}
-
-function openRiskHistoryItem(index) {
-  const item = riskAnalysisHistory[index];
-  if (!item) return;
-  renderRiskSnapshot(item.snapshot || null);
-  renderRiskStructured(item.structured || null);
-  document.getElementById('riskResult').textContent = item.content || '历史记录无内容。';
-  const usageText = formatRiskUsage(item.usage || null);
-  applyRiskStatus(usageText ? '已打开历史分析。' + usageText : '已打开历史分析。', '');
-}
-
-function openRiskTimelineFromHistory(index) {
-  const item = riskAnalysisHistory[index];
-  if (!item) return;
-  openEventTimelineAround(item.analysis_time || (item.snapshot && item.snapshot.analysis_time), 'risk_analysis', item.id);
-}
-
-function handleRiskHistoryTimelineClick(event) {
-  const button = event.target && event.target.closest ? event.target.closest('[data-risk-timeline-index]') : null;
-  if (!button) return;
-  event.preventDefault();
-  event.stopPropagation();
-  openRiskTimelineFromHistory(Number(button.getAttribute('data-risk-timeline-index')));
-}
-
-document.addEventListener('click', handleRiskHistoryTimelineClick, true);
-
-function clearRiskHistory() {
-  if (riskAnalysisRunning) return;
-  socket.emit('clear_risk_analysis_history');
-}
-
-function currentRiskReportMarkdown() {
-  const result = document.getElementById('riskResult').textContent || '';
-  const meta = Array.from(document.querySelectorAll('#riskMeta span')).map(item => item.textContent).join('\n');
-  const evidence = document.getElementById('riskEvidence') ? document.getElementById('riskEvidence').innerText.trim() : '';
-  const parts = ['# 风险分析报告'];
-  if (meta) parts.push('', '## 数据快照', meta);
-  if (evidence) parts.push('', '## 数据依据', evidence);
-  parts.push('', '## 分析内容', result || '暂无分析结果。');
-  return parts.join('\n');
-}
-
-function copyRiskReport() {
-  const content = currentRiskReportMarkdown();
-  if (navigator.clipboard && navigator.clipboard.writeText) {
-    navigator.clipboard.writeText(content).then(() => applyRiskStatus('报告已复制。', '')).catch(() => applyRiskStatus('复制失败，请手动选择内容。', 'error'));
-    return;
-  }
-  applyRiskStatus('当前环境不支持自动复制，请手动选择内容。', 'error');
-}
-
-function exportRiskReport() {
-  const stamp = new Date().toISOString().replace(/[:.]/g, '-');
-  downloadText('GoldMonitor-risk-report-' + stamp + '.md', currentRiskReportMarkdown(), 'text/markdown;charset=utf-8');
-  applyRiskStatus('报告已导出。', '');
-}
 
 // ========== 更新 ==========
 function openUpdate() {
@@ -5266,7 +3898,6 @@ function clearAlertLog() {
   socket.emit('clear_alert_log');
 }
 
-window.openEventTimelineAround = openEventTimelineAround;
 window.openAlertTimelineFromLog = openAlertTimelineFromLog;
 window.openRiskTimelineFromHistory = openRiskTimelineFromHistory;
